requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
dataclasses-json==0.6.5
diskcache==5.6.3
dnspython==2.6.1
exceptiongroup==1.2.1
fastapi==0.110.3
fastparquet==2024.2.0
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.23.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.52
llama_cpp_python==0.2.67
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
psutil==5.9.8
pyarrow==16.1.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.6.3
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tenacity==8.2.3
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
transformers==4.40.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==12.0
yarl==1.9.4
